Adding profile settings
