+++ lib/config/routes/route_handlers.dart
@@ -1,4 +1,5 @@
 import 'package:demos_app/modules/settings/screens/general_settings.screen.dart';
+import 'package:demos_app/modules/settings/screens/profile.screen.dart';
 import 'package:demos_app/modules/spaces/screens/spaces.screen.dart';
 import 'package:flutter/material.dart';
 import 'package:fluro/fluro.dart';
@@ -30,4 +31,9 @@ var spacesHandler = Handler(
 var generalSettingsHandler = Handler(
     handlerFunc: (BuildContext? context, Map<String, List<String>> params) {
   return GeneralSettingsScreen();
+});
+
+var profileSettingsHandler = Handler(
+    handlerFunc: (BuildContext? context, Map<String, List<String>> params) {
+  return ProfileSettingScreen();
 });
