+++ lib/modules/settings/screens/profile.screen.dart
@@ -1,17 +1,21 @@
+import 'package:demos_app/widgets/profile/profile.widget.dart';
 import 'package:flutter/material.dart';
 
-class Profile extends StatefulWidget {
-  Profile({Key? key}) : super(key: key);
+class ProfileSettingScreen extends StatefulWidget {
+  ProfileSettingScreen({Key? key}) : super(key: key);
 
   @override
   _ProfileState createState() => _ProfileState();
 }
 
-class _ProfileState extends State<Profile> {
+class _ProfileState extends State<ProfileSettingScreen> {
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: Text('working'),
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Perfil'),
+      ),
+      body: Profile(),
     );
   }
 }
